# internal/api/handlers.go
// SPDX-FileCopyrightText: (C) 2026 Intel Corporation
// SPDX-License-Identifier: Apache-2.0

package api

import (
	"encoding/json"
	"errors"
	"net/http"
	"time"

	httpapi "github.com/open-edge-platform/image-composer-tool/internal/api/http"
	"github.com/open-edge-platform/image-composer-tool/internal/api/service"
)

// Server implements the generated httpapi.ServerInterface. Each method decodes
// the request into the generated types, calls the service, maps the result back
// to the generated response type, and encodes it — no business logic here.
var _ httpapi.ServerInterface = (*Server)(nil)

// GetManifest handles GET /manifest.
func (s *Server) GetManifest(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, fromManifest(s.svc.Manifest()))
}

// ComposeTemplate handles POST /templates/compose.
func (s *Server) ComposeTemplate(w http.ResponseWriter, r *http.Request) {
	var req httpapi.ComposeRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "BAD_REQUEST", "invalid JSON body")
		return
	}
	res, err := s.svc.Compose(toSelection(req))
	if err != nil {
		writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, fromComposeResult(res))
}

// StartBuild handles POST /builds.
func (s *Server) StartBuild(w http.ResponseWriter, r *http.Request) {
	var req httpapi.BuildRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "BAD_REQUEST", "invalid JSON body")
		return
	}
	acc, err := s.svc.StartBuild(toBuildRequest(req))
	if err != nil {
		writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusAccepted, fromBuildAccepted(acc))
}

// CancelBuild handles POST /builds/{id}/cancel.
func (s *Server) CancelBuild(w http.ResponseWriter, _ *http.Request, id httpapi.BuildId) {
	acc, err := s.svc.CancelBuild(id)
	if err != nil {
		writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusAccepted, fromCancelAccepted(acc))
}

// ListBuilds handles GET /builds.
func (s *Server) ListBuilds(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, fromHistory(s.svc.ListBuilds(), time.Now().UTC()))
}

// GetBuildDetails handles GET /builds/{id}/details.
func (s *Server) GetBuildDetails(w http.ResponseWriter, _ *http.Request, id httpapi.BuildId) {
	d, err := s.svc.BuildDetails(id)
	if err != nil {
		writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, fromBuildDetails(d))
}

// ListBuildArtifacts handles GET /builds/{id}/artifacts.
func (s *Server) ListBuildArtifacts(w http.ResponseWriter, _ *http.Request, id httpapi.BuildId) {
	l, err := s.svc.BuildArtifacts(id)
	if err != nil {
		writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, fromArtifactList(l))
}

// ListPackageRepos handles GET /package-repos.
//
// The optional `os` query parameter filters the catalog to repositories offered
// for that target; omitting it returns the whole catalog. An unknown target is
// not an error — it simply has no repositories, which the UI renders as an empty
// picker.
func (s *Server) ListPackageRepos(w http.ResponseWriter, _ *http.Request, params httpapi.ListPackageReposParams) {
	var osID string
	if params.Os != nil {
		osID = *params.Os
	}
	writeJSON(w, http.StatusOK, fromPackageRepoList(s.svc.PackageRepos(osID)))
}

// SearchPackages handles GET /packages/search.
//
// Omitting `q` browses a repository's catalog instead of searching it — the
// `limit`/`offset` pair pages through the result either way.
func (s *Server) SearchPackages(w http.ResponseWriter, r *http.Request, params httpapi.SearchPackagesParams) {
	var query string
	if params.Q != nil {
		query = *params.Q
	}
	var repos []string
	if params.Repos != nil {
		repos = *params.Repos
	}
	limit := 0
	if params.Limit != nil {
		limit = *params.Limit
	}
	offset := 0
	if params.Offset != nil {
		offset = *params.Offset
	}
	hits, total, err := s.svc.SearchPackages(r.Context(), params.Os, query, repos, limit, offset)
	if err != nil {
		writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, fromPackageSearchResults(query, hits, total))
}

// ValidateTemplate handles POST /templates/validate.
//
// A failed validation is a successful call: invalid templates return 200 with
// valid=false and one issue per problem. Only a malformed request body or an
// empty template yields a 4xx.
func (s *Server) ValidateTemplate(w http.ResponseWriter, r *http.Request) {
	var req httpapi.ValidateTemplateJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "BAD_REQUEST", "invalid JSON body")
		return
	}
	res, err := s.svc.ValidateTemplate(req.Yaml)
	if err != nil {
		writeServiceError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, fromValidationResult(res))
}

// writeServiceError maps a service error onto the JSON error envelope. Domain
// *service.Error carry an HTTP status and machine code; any other error is an
// opaque 500 so internal details don't leak to clients.
func writeServiceError(w http.ResponseWriter, err error) {
	var se *service.Error
	if errors.As(err, &se) {
		writeError(w, se.Status, se.Code, se.Message)
		return
	}
	writeError(w, http.StatusInternalServerError, "INTERNAL", "internal server error")
}